refactor: Extracted generic page rendering into pageController
