+++ lib/handlers.js
@@ -1,27 +1,14 @@
-const db = require('../database/database')
+const db = require('../src/config/database')
 const path = require('path')
 const bcrypt = require("bcrypt")
 const sharp = require('sharp')
 const fs = require('fs')
 exports.api = {}
 
-exports.home = (req, res) => {
-    res.render('home')
-}
-
 exports.cadastrarPratos = (req, res) => {
     res.render('novosPratos',{ user: req.session.user, rota: req.params.username} )
 }
 
-
-
-exports.restaurantes = (req, res) => {
-    let stmt = db.prepare(`
-        SELECT nome, username
-        FROM users`)
-    let result = stmt.all()
-    res.render('restaurantes', { result })
-}
 
 exports.api.cadastrarPratos = (req, res, fields, files) => {
     let now = `_${Date.now()}`
